[Platinum II] Title: 잔디밭의 개미굴, Time: 280 ms, Memory: 42348 KB, Score: 100 point -BaekjoonHub

+++ 백준/Platinum/28328. 잔디밭의 개미굴/잔디밭의 개미굴.cc
@@ -4,7 +4,7 @@ using namespace std;
 using ll = long long;
 using pii = pair<int, int>;
 
-int N, M;
+ll N, M;
 vector<int> tree[250005];
 int dp[250005][2], vst[250005][2];
